backend/pyproject.toml
[tool.poetry]
name = "society-backend"
version = "0.1.0"
description = "Backend API for Girchi Digital Polis"
authors = ["Girchi Digital Polis Team"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
pydantic = {extras = ["email"], version = "^2.5.1"}
pydantic-settings = "^2.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
asyncpg = "^0.29.0"
redis = "^5.0.1"
celery = "^5.3.6"
httpx = "^0.25.1"
bcrypt = "^4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
